Using height textures to implement shadows.
Shadows are created only by a simple check on
height maps. This has to be replaced when
deferred rendering is used.

It also need to be more then just a check, wether
or not a given texel is above the current texel,
but take into account, how far that texel is, and
in what direction the light source is.